added dp83867 clk output dt property
The dp83867 Linux PHY driver issues a soft reset upon init that will
unconfigure things done in the boot firmware critical to operation such
as rgmii internal delays, LED configuration, and clk output.

Support for configuring the clk output was added in
9708fb630d19ee51ae3aeb3a533e3010da0e8570 which will at least get the phy
working for Newport boards using this RGMII PHY.

It is still not recommended to use this phy driver on Newport boards as it
will undo the LED configuration done in the boot firmware.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>